refactor: cash mantissa

+++ subgraphs/isolated-pools/src/utilities/exponentToBigInt.ts
@@ -1,6 +1,6 @@
 import { BigInt } from '@graphprotocol/graph-ts';
 
-function exponentToBigInt(decimals: i32): i32 {
+function exponentToBigInt(decimals: i32): BigInt {
   let bd = BigInt.fromString('1');
   for (let i = 0; i < decimals; i++) {
     bd = bd.times(BigInt.fromString('10'));
