+++ subgraphs/venus/src/utilities/getUnderlyingPrice.ts
@@ -1,8 +1,8 @@
-import { Address } from '@graphprotocol/graph-ts';
+import { Address, BigInt } from '@graphprotocol/graph-ts';
 
 import { getTokenPriceCents } from './getTokenPriceCents';
 
-export function getUnderlyingPrice(address: string, underlyingDecimals: i32): i32 {
+export function getUnderlyingPrice(address: string, underlyingDecimals: i32): BigInt {
   const contractAddress = Address.fromString(address);
   const underlyingPriceCents = getTokenPriceCents(contractAddress, underlyingDecimals);
